# apps/backend/docs/swagger.yaml
basePath: /
definitions:
...
  CreateRequest:
    properties:
      name:
        minLength: 1
        type: string
      slug:
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
...
  ErrorResponse:
    properties:
      message:
        type: string
    type: object
  InviteRequest:
    properties:
      email:
        type: string
      role:
        enum:
        - admin
        - member
        type: string
    required:
    - email
    - role
    type: object
...
  LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  LogoutResponse:
    properties:
      success:
        type: boolean
    type: object
  MembershipResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      organization:
        $ref: '#/definitions/OrgResponse'
      organizationId:
        type: string
      role:
        type: string
      userId:
        type: string
    type: object
  OrgResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  RegisterRequest:
    properties:
      displayName:
        minLength: 1
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
  TokenResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
info:
  contact: {}
  description: Multi-tenant B2B SaaS control-plane API (Go port of controlplane-api).
  title: Controlplane API
  version: 0.1.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "401":
          description: INVALID_CREDENTIALS
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: TOO_MANY_ATTEMPTS
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <accessToken>
        in: header
        name: Authorization
        type: string
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LogoutResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "401":
          description: INVALID_REFRESH_TOKEN / REFRESH_TOKEN_REUSE / REFRESH_TOKEN_EXPIRED
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Rotate a refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TokenResponse'
        "409":
          description: EMAIL_TAKEN
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
  /organizations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MembershipResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List my organizations
      tags:
      - organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/OrgResponse'
        "409":
          description: SLUG_TAKEN
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create an organization
      tags:
      - organizations
  /organizations/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Invite payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/InviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: FORBIDDEN / LIMIT_EXCEEDED / Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: ALREADY_MEMBER
          schema:
            $ref: '#/definitions/ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite a member
      tags:
      - organizations
  /organizations/members/{userId}:
    delete:
      parameters:
      - description: Active organization ID
        in: header
        name: x-organization-id
        required: true
        type: string
      - description: Member user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Missing x-organization-id header
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: CANNOT_REMOVE_OWNER / FORBIDDEN / Not a member of this organization
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: MEMBER_NOT_FOUND
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a member
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" — the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"